#4 change balance value type to double

+++ src/main/kotlin/pl/szymonleyk/betgame/register/account/Account.kt
@@ -4,4 +4,10 @@ import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Table
 
 @Table
-data class Account(@Id val id: Int? = null, val username: String, val name: String, val surname: String, val balance: Long)
+data class Account(
+    @Id val id: Int? = null,
+    val username: String,
+    val name: String,
+    val surname: String,
+    val balance: Double
+)
